Create automated_DEA
An automated code that initially allows you to navigate in ArrayExpress to select the desired datasets that match your analysis interest. Once the data are selected, it performs a differential expression analysis, identifying the platform (Affymetrix) and performing the correspondent statistical inference and treatment (Limma). 

Note: The code is still under construction. It is also required to make it able to navigate in the new repository called "Biostudies".