Create HMM_decoder_and_evaluator
Using an HMM specified  by the user (transition_df & emission_df) calculates for a particular sequence (seq) its probability (action="eval") or the most probable path that would generate it (action="decode"). The strategy used by the algorithm is simmilar to the one employed by the forward algorithm, however, it does not use the Viterbi strategy. The function calculates all the potential paths, thus, it is not optimised and it is not still recomendable for very long sequences or models with silent states.